--- specification/xpack/usage/DataTiersUsage.ts
@@ -1,42 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { long } from '@_types/Numeric'
-import { XPackUsage } from './XPackUsage'
-
-export class DataTiersUsage extends XPackUsage {
-  data_warm: DataTierPhaseCountUsage
-  data_frozen?: DataTierPhaseCountUsage // @since 7.13
-  data_cold: DataTierPhaseCountUsage
-  data_content: DataTierPhaseCountUsage
-  data_hot: DataTierPhaseCountUsage
-}
-
-export class DataTierPhaseCountUsage {
-  node_count: long
-  index_count: long
-  total_shard_count: long
-  primary_shard_count: long
-  doc_count: long
-  total_size_bytes: long
-  primary_size_bytes: long
-  primary_shard_size_avg_bytes: long
-  primary_shard_size_median_bytes: long
-  primary_shard_size_mad_bytes: long
-}
