--- specification/xpack/usage/RuntimeFieldsUsage.ts
@@ -1,43 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { Field } from '@_types/common'
-import { long } from '@_types/Numeric'
-import { XPackUsage } from './XPackUsage'
-
-export class RuntimeFieldsUsage extends XPackUsage {
-  field_types: RuntimeFieldsTypeUsage[]
-}
-
-export class RuntimeFieldsTypeUsage {
-  chars_max: long
-  chars_total: long
-  count: long
-  doc_max: long
-  doc_total: long
-  index_count: long
-  lang: string[]
-  lines_max: long
-  lines_total: long
-  name: Field
-  scriptless_count: long
-  shadowed_count: long
-  source_max: long
-  source_total: long
-}
