--- specification/xpack/usage/SecurityRolesUsage.ts
@@ -1,49 +1,0 @@
-/*
- * Licensed to Elasticsearch B.V. under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch B.V. licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-import { ByteSize } from '@_types/common'
-import { integer, long, ulong } from '@_types/Numeric'
-
-export class SecurityRolesUsage {
-  native: SecurityRolesNativeUsage
-  dls: SecurityRolesDlsUsage
-  file: SecurityRolesFileUsage
-}
-
-export class SecurityRolesNativeUsage {
-  dls: boolean
-  fls: boolean
-  size: long
-}
-
-export class SecurityRolesDlsUsage {
-  bit_set_cache: SecurityRolesDlsBitSetCacheUsage
-}
-
-export class SecurityRolesDlsBitSetCacheUsage {
-  count: integer
-  memory: ByteSize
-  memory_in_bytes: ulong
-}
-
-export class SecurityRolesFileUsage {
-  dls: boolean
-  fls: boolean
-  size: long
-}
